Allow the WebDriverConfig class handle browserPreferencess
We want to be able to include browserPreferences as a key in the root of the config file
and have the resulting WebDriverConfig object expose methods to fetch specific properties
based on the provided browser type. This can then be picked up when the drivers are created.

This commit only lets the WebDriverCongfig class support this extra field.
Following commits will have the code that includes the configs for the supported driver types.

In terms of testing, we have not included fixture files for the various test cases (as is the case in the other tests)
So we should probably have a chat about that and refactor the code to either use the fixture files or leave as-is.

The reason  we choose not to crerate the extra fixture files is that it would require at least 3 extra files for the 4 test cases
wich can be a bit excessive especially as defining the input inline does not make the code less readable?
Will be happy to change this approach if others think otherwise.

We also handled the scenario where the case of the browser type does not matter.
This can also be removed if deemed unnecessary.